+++ src/bot.js
@@ -34,4 +34,32 @@ client.on('message', (message) => {
     }
 })
 
+client.on('presenceUpdate', (oldMember, newMember) => {
+    // only doing something if the person has come online
+    if (oldMember.presence.status != 'offline') return;
+
+    const channel = getDefaultChannel(newMember.guild);
+
+    channel.send(`Welcome ${newMember} to the server!`);
+});
+
+// create function to find default or first text channel
+const getDefaultChannel = (guild) => {
+    // get "original" default channel
+    if (guild.channels.has(guild.id)) return guild.channels.get(guild.id)
+
+    // Check for a "general" channel, which is often default chat
+    const generalChannel = guild.channels.find(channel => channel.name === "general");
+
+    if (generalChannel) return generalChannel;
+    // Now we get into the heavy stuff: first channel in order where the bot can speak
+    // hold on to your hats!
+    return guild.channels
+        .filter(c => c.type === "text" &&
+            c.permissionsFor(guild.client.user).has("SEND_MESSAGES"))
+        .sort((a, b) => a.position - b.position ||
+            Long.fromString(a.id).sub(Long.fromString(b.id)).toNumber())
+        .first();
+}
+
 client.login(token);